requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
cloudscraper==1.2.68
idna==3.4
markdown-it-py==2.1.0
mdurl==0.1.2
Pillow==9.4.0
PixivPy3==3.7.2
Pygments==2.14.0
pyparsing==3.0.9
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.1
typing-extensions==4.5.0
urllib3==1.26.14
